Move ExecutionContext to execution_context.hpp
