+++ lib/fizzy/execute.hpp
@@ -6,6 +6,7 @@
 
 #include "cxx20/span.hpp"
 #include "exceptions.hpp"
+#include "execution_context.hpp"
 #include "instantiate.hpp"
 #include "limits.hpp"
 #include "module.hpp"
@@ -41,33 +42,6 @@ constexpr ExecutionResult Void{true};
 /// Shortcut for execution that resulted in a trap.
 constexpr ExecutionResult Trap{false};
 
-/// The storage for information shared by calls in the same execution "thread".
-/// Users may decide how to allocate the execution context, but some good defaults are available.
-class ExecutionContext
-{
-    /// Call depth increment guard.
-    /// It will automatically decrement the call depth to the original value
-    /// when going out of scope.
-    class [[nodiscard]] Guard
-    {
-        ExecutionContext& m_execution_context;  ///< Reference to the guarded execution context.
-
-    public:
-        explicit Guard(ExecutionContext& ctx) noexcept : m_execution_context{ctx} {}
-        ~Guard() noexcept { --m_execution_context.depth; }
-    };
-
-public:
-    int depth = 0;  ///< Current call depth.
-
-    /// Increments the call depth and returns the guard object which decrements
-    /// the call depth back to the original value when going out of scope.
-    Guard increment_call_depth() noexcept
-    {
-        ++depth;
-        return Guard{*this};
-    }
-};
 
 /// Execute a function from an instance.
 ///
